requirements.txt
boto3==1.26.13
dj-database-url==1.0.0
Django==3.2.16
django-bmemcached==0.3.0
django-markdown==0.8.4
django-pagedown==2.2.1
django-storages==1.13.1
gunicorn==20.1.0 
Markdown==3.4.1
Pillow==9.3.0
psycopg2-binary==2.9.5
Pygments==2.13.0
